AS Unit Tests - Truezip wrong classloader
