+++ as/plugins/org.jboss.ide.eclipse.as.wtp.core/src/org/jboss/ide/eclipse/as/wtp/core/server/publish/LocalZippedModulePublishRunner.java
@@ -33,8 +33,7 @@ import org.eclipse.wst.server.core.model.IModuleFile;
 import org.eclipse.wst.server.core.model.IModuleFolder;
 import org.eclipse.wst.server.core.model.IModuleResource;
 import org.eclipse.wst.server.core.model.IModuleResourceDelta;
-import org.jboss.ide.eclipse.archives.core.util.internal.TrueZipUtil;
-import org.jboss.ide.eclipse.archives.core.util.internal.TrueZipUtil.JarArchiveDetector;
+import org.jboss.ide.eclipse.archives.core.util.TrueZipUtil;
 import org.jboss.ide.eclipse.as.core.server.IModulePathFilter;
 import org.jboss.ide.eclipse.as.core.server.IModulePathFilterProvider;
 import org.jboss.ide.eclipse.as.core.util.FileUtil;
@@ -46,7 +45,7 @@ import org.jboss.ide.eclipse.as.wtp.core.Messages;
 import org.jboss.ide.eclipse.as.wtp.core.server.behavior.util.PublishControllerUtil;
 import org.jboss.ide.eclipse.as.wtp.core.util.ServerModelUtilities;
 
-import de.schlichtherle.io.ArchiveDetector;
+
 
 /**
  * This class is a utility class meant to zip a given module
@@ -195,7 +194,7 @@ public class LocalZippedModulePublishRunner extends ModuleResourceUtil {
 				// We're the root module (no parent) so just create the archive to start
 				moduleRoot = TrueZipUtil.getFile(moduleAbsoluteDestination, TrueZipUtil.getJarArchiveDetector());
 				// Clear out existing, in case some child modules were removed
-				if( moduleRoot.exists()) {
+				if( TrueZipUtil.pathExists(moduleRoot)) {
 					TrueZipUtil.deleteAll(moduleRoot);
 				}
 				boolean createWorked = TrueZipUtil.createArchive(moduleAbsoluteDestination);
@@ -252,7 +251,7 @@ public class LocalZippedModulePublishRunner extends ModuleResourceUtil {
 		try {
 			IPath tail = getRootModuleRelativePath(combine(parent, last));
 			IPath tailLocation = tail.removeLastSegments(1);
-			de.schlichtherle.io.File root = TrueZipUtil.getFile(destinationArchive.append(tailLocation), TrueZipUtil.getJarArchiveDetector());
+			java.io.File root = TrueZipUtil.getFile(destinationArchive.append(tailLocation), TrueZipUtil.getJarArchiveDetector());
 			IModuleResource[] resources = getResources(last, new NullProgressMonitor());
 			int total = countMembers(resources, true);
 			monitor.beginTask("Copying Resources", total*100);
@@ -363,7 +362,7 @@ public class LocalZippedModulePublishRunner extends ModuleResourceUtil {
 
 	private IStatus[] publishChanges(IModule[] module, IProgressMonitor monitor) {
 		IPath path = destinationArchive.append(getRootModuleRelativePath(module));
-		de.schlichtherle.io.File root = TrueZipUtil.getFile(path, TrueZipUtil.getJarArchiveDetector());
+		java.io.File root = TrueZipUtil.getFile(path, TrueZipUtil.getJarArchiveDetector());
 		IModuleResourceDelta[] deltas = getDeltaForModule(module);
 		IModulePathFilter filter = filterProvider == null ? null : filterProvider.getFilter(server, module);
 		return publishChanges(deltas, root, filter, monitor);
@@ -389,7 +388,7 @@ public class LocalZippedModulePublishRunner extends ModuleResourceUtil {
 	
 	
 	private IStatus[] publishChanges(IModuleResourceDelta[] deltas, 
-			de.schlichtherle.io.File root, IModulePathFilter filter, IProgressMonitor monitor) {
+			java.io.File root, IModulePathFilter filter, IProgressMonitor monitor) {
 		ArrayList<IStatus> results = new ArrayList<IStatus>();
 		if( deltas == null || deltas.length == 0 )
 			return new IStatus[]{};
@@ -409,10 +408,10 @@ public class LocalZippedModulePublishRunner extends ModuleResourceUtil {
 					results.addAll(Arrays.asList(publishChanges(deltas[i].getAffectedChildren(), root, filter, monitor)));
 				}
 			} else if( dKind == IModuleResourceDelta.REMOVED) {
-				de.schlichtherle.io.File f = getDestinationJar(root, 
+				java.io.File f = getDestinationJar(root, 
 						resource.getModuleRelativePath().append(
 								resource.getName()));
-				boolean b = f.deleteAll();
+				boolean b = TrueZipUtil.deleteAll(f);
 				if( !b )
 					results.add(generateDeleteFailedStatus(f));
 			} else if( dKind == IModuleResourceDelta.NO_CHANGE  ) {
@@ -424,21 +423,21 @@ public class LocalZippedModulePublishRunner extends ModuleResourceUtil {
 	}
 
 	
-	private IStatus[] copy(de.schlichtherle.io.File root, IModuleResource[] children, IProgressMonitor monitor) {
+	private IStatus[] copy(java.io.File root, IModuleResource[] children, IProgressMonitor monitor) {
 		ArrayList<IStatus> results = new ArrayList<IStatus>();
 		for( int i = 0; i < children.length; i++ ) {
 			if( children[i] instanceof IModuleFile ) {
 				IModuleFile mf = (IModuleFile)children[i];
 				java.io.File source = getFile(mf);
 				if( source != null ) {
-					de.schlichtherle.io.File destination = getDestinationJar(root, mf.getModuleRelativePath().append(mf.getName()));
-					boolean b = new de.schlichtherle.io.File(source, ArchiveDetector.NULL).archiveCopyAllTo(destination);
+					java.io.File destination = getDestinationJar(root, mf.getModuleRelativePath().append(mf.getName()));
+					boolean b = TrueZipUtil.archiveCopyAllTo(source, TrueZipUtil.getNullArchiveDetector(), destination);
 					if( !b )
 						results.add(generateCopyFailStatus(source, destination));
 				}
 				monitor.worked(100);
 			} else if( children[i] instanceof IModuleFolder ) {
-				de.schlichtherle.io.File destination = getDestinationJar(root, children[i].getModuleRelativePath().append(children[i].getName()));
+				java.io.File destination = getDestinationJar(root, children[i].getModuleRelativePath().append(children[i].getName()));
 				destination.mkdirs();
 				IModuleFolder mf = (IModuleFolder)children[i];
 				results.addAll(Arrays.asList(copy(root, mf.members(), monitor)));
@@ -457,18 +456,8 @@ public class LocalZippedModulePublishRunner extends ModuleResourceUtil {
 		return new Status(IStatus.ERROR, ASWTPToolsPlugin.PLUGIN_ID, "Copy of " + source + " to " + destination + " has failed");//$NON-NLS-1$//$NON-NLS-2$//$NON-NLS-3$
 	}
 	
-	private de.schlichtherle.io.File getDestinationJar(de.schlichtherle.io.File root, IPath relative) {
-		while(relative.segmentCount() > 0 ) {
-			if( relative.segmentCount() == 1 ) {
-				root = new de.schlichtherle.io.File(root, 
-						relative.segment(0), JarArchiveDetector.ALL);
-			} else {
-				root = new de.schlichtherle.io.File(root, 
-						relative.segment(0), ArchiveDetector.NULL);
-			}
-			relative = relative.removeFirstSegments(1);
-		}
-		return root;
+	private java.io.File getDestinationJar(java.io.File root, IPath relative) {
+		return TrueZipUtil.getDestinationJar(root, relative);
 	}
 	
 	private de.schlichtherle.io.File getFileInArchive(de.schlichtherle.io.File root, IPath relative) {
